+++ JWT/DTO/ExamDto/QuestionDto.cs
@@ -3,7 +3,7 @@
 	public class QuestionDto
 	{
 		public string QuestionText { get; set; }  
-		public int Marks { get; set; }  
+		public double Marks { get; set; }  
 
 		public int TimeInMin { get; set; }
 		public List<ChoiceDto> Choices { get; set; }
