+++ JWT/Model/Exams/ExamStudent.cs
@@ -9,7 +9,7 @@
 		public int ExamId { get; set; }
 		public Exam Exam { get; set; }
 
-		public int Score { get; set; } // درجة الطالب في الامتحان
+		public double Score { get; set; } // درجة الطالب في الامتحان
 	   public bool IsAbsent { get; set; } = true;
 	   public int precentageExam { get; set; }
 	}
